# openapi.yaml
openapi: 3.1.0
info:
  title: RSS Subscription Summarizer API
  description: API for managing RSS feeds, crawling articles, and generating LLM-powered summaries.
  version: 1.0.0

servers:
  - url: http://localhost:8000/api/v1
    description: Local development server

tags:
  - name: Sources
    description: RSS feed source management (CRUD)
  - name: Articles
    description: Fetched articles (read-only)
  - name: Summaries
    description: LLM-generated summaries
  - name: Crawler
    description: Manual and scheduled RSS crawling
  - name: Scheduler
    description: Automatic crawl scheduler configuration
  - name: Settings
    description: Application settings (LLM provider, etc.)

paths:
  # ─── Sources ───────────────────────────────────────────
  /sources:
    get:
      tags: [Sources]
      summary: List all RSS sources
      operationId: listSources
      responses:
        "200":
          description: List of RSS sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Source"
    post:
      tags: [Sources]
      summary: Add a new RSS source
      operationId: createSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceCreate"
      responses:
        "201":
          description: Source created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "422":
          $ref: "#/components/responses/ValidationError"

  /sources/{source_id}:
    parameters:
      - $ref: "#/components/parameters/SourceId"
    get:
      tags: [Sources]
      summary: Get a single RSS source
      operationId: getSource
      responses:
        "200":
          description: RSS source details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Sources]
      summary: Update an RSS source
      operationId: updateSource
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceUpdate"
      responses:
        "200":
          description: Source updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Source"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [Sources]
      summary: Delete an RSS source
      operationId: deleteSource
      responses:
        "204":
          description: Source deleted
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Articles ──────────────────────────────────────────
  /articles:
    get:
      tags: [Articles]
      summary: List fetched articles
      operationId: listArticles
      parameters:
        - name: source_id
          in: query
          description: Filter by source ID
          schema:
            type: string
            format: uuid
        - name: since
          in: query
          description: Only articles published after this datetime (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Max number of articles to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of articles to skip
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"

  /articles/{article_id}:
    parameters:
      - name: article_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Articles]
      summary: Get a single article
      operationId: getArticle
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Summaries ─────────────────────────────────────────
  /summaries:
    get:
      tags: [Summaries]
      summary: List generated summaries
      operationId: listSummaries
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryListResponse"

  /summaries/{summary_id}:
    parameters:
      - name: summary_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Summaries]
      summary: Get a single summary
      operationId: getSummary
      responses:
        "200":
          description: Summary details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Summaries]
      summary: Delete a summary
      operationId: deleteSummary
      responses:
        "204":
          description: Summary deleted
        "404":
          $ref: "#/components/responses/NotFound"

  /summaries/generate:
    post:
      tags: [Summaries]
      summary: Generate a new summary
      description: |
        Manually trigger LLM summarization. Supports two modes:
        - **All sources (past N days)**: omit `source_ids` or pass empty array
        - **Specific sources**: pass one or more `source_ids`
      operationId: generateSummary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SummaryGenerateRequest"
      responses:
        "200":
          description: Summary generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        "400":
          description: No articles found for the given criteria
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
        "502":
          description: LLM provider error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Crawler ───────────────────────────────────────────
  /crawler/trigger:
    post:
      tags: [Crawler]
      summary: Trigger an immediate crawl
      description: Crawls all enabled RSS sources. Only fetches articles — does NOT generate summaries.
      operationId: triggerCrawl
      responses:
        "200":
          description: Crawl completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrawlResult"

  # ─── Scheduler ─────────────────────────────────────────
  /scheduler/status:
    get:
      tags: [Scheduler]
      summary: Get scheduler status
      operationId: getSchedulerStatus
      responses:
        "200":
          description: Current scheduler status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerStatus"

  /scheduler/config:
    put:
      tags: [Scheduler]
      summary: Update scheduler configuration
      description: Update auto-crawl interval. The scheduler only handles crawling — summarization is always manual.
      operationId: updateSchedulerConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchedulerConfigUpdate"
      responses:
        "200":
          description: Scheduler config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerStatus"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ─── Settings ──────────────────────────────────────────
  /settings:
    get:
      tags: [Settings]
      summary: Get application settings
      operationId: getSettings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      tags: [Settings]
      summary: Update application settings
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "422":
          $ref: "#/components/responses/ValidationError"

# ═══════════════════════════════════════════════════════════
# Components
# ═══════════════════════════════════════════════════════════
components:
  parameters:
    SourceId:
      name: source_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    # ─── Source ────────────────────────────────────────────
    Source:
      type: object
      required: [id, name, url, category, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "Hacker News"
        url:
          type: string
          format: uri
          example: "https://news.ycombinator.com/rss"
        category:
          type: string
          example: "tech"
        description:
          type: string
          nullable: true
          example: "Turing Award winner and 'Godfather of Deep Learning'."
        enabled:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_fetched_at:
          type: string
          format: date-time
          nullable: true

    SourceCreate:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: "Hacker News"
        url:
          type: string
          format: uri
          example: "https://news.ycombinator.com/rss"
        category:
          type: string
          default: "general"
          example: "tech"
        description:
          type: string
          example: "A popular tech news aggregator."
        enabled:
          type: boolean
          default: true

    SourceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        url:
          type: string
          format: uri
        category:
          type: string
        description:
          type: string
        enabled:
          type: boolean

    # ─── Article ───────────────────────────────────────────
    Article:
      type: object
      required: [id, source_id, title, url, fetched_at, guid]
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        title:
          type: string
          example: "Understanding Transformer Architectures"
        url:
          type: string
          format: uri
        author:
          type: string
          nullable: true
          example: "John Doe"
        published_at:
          type: string
          format: date-time
          nullable: true
        fetched_at:
          type: string
          format: date-time
        content_snippet:
          type: string
          description: First 500 characters of the article content
          example: "This article explores..."
        guid:
          type: string
          description: Unique RSS entry identifier (used for deduplication)

    ArticleListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
          description: Total number of matching articles
          example: 142
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0

    # ─── Summary ───────────────────────────────────────────
    Summary:
      type: object
      required: [id, created_at, period_start, period_end, source_ids, article_count, llm_provider, llm_model, content]
      properties:
        id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        period_start:
          type: string
          format: date-time
        period_end:
          type: string
          format: date-time
        source_ids:
          type: array
          items:
            type: string
            format: uuid
          description: IDs of sources included in this summary
        article_count:
          type: integer
          example: 42
        llm_provider:
          type: string
          enum: [claude, openai]
        llm_model:
          type: string
          example: "claude-sonnet-4-20250514"
        content:
          type: string
          description: Markdown-formatted summary content
          example: "## Trends\n\n### AI/ML\n..."
        prompt_tokens:
          type: integer
          example: 8500
        completion_tokens:
          type: integer
          example: 1200

    SummaryListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Summary"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    SummaryGenerateRequest:
      type: object
      properties:
        source_ids:
          type: array
          items:
            type: string
            format: uuid
          description: IDs of sources to include. Empty or omitted = all enabled sources.
          default: []
        since_days:
          type: integer
          description: Include articles from the past N days
          default: 7
          minimum: 1
          maximum: 90

    # ─── Crawler ───────────────────────────────────────────
    CrawlResult:
      type: object
      required: [total_new_articles, sources_crawled, errors]
      properties:
        total_new_articles:
          type: integer
          description: Number of new (deduplicated) articles fetched
          example: 23
        sources_crawled:
          type: integer
          description: Number of sources successfully crawled
          example: 5
        errors:
          type: array
          items:
            $ref: "#/components/schemas/CrawlError"
          description: Errors encountered during crawling (partial failures)

    CrawlError:
      type: object
      required: [source_id, source_name, error]
      properties:
        source_id:
          type: string
          format: uuid
        source_name:
          type: string
        error:
          type: string
          example: "Connection timeout after 30s"

    # ─── Scheduler ─────────────────────────────────────────
    SchedulerStatus:
      type: object
      required: [enabled, interval_minutes, max_articles_per_summary, running]
      properties:
        enabled:
          type: boolean
          example: true
        interval_minutes:
          type: integer
          example: 360
        max_articles_per_summary:
          type: integer
          description: Max articles to include when generating a summary
          example: 100
        running:
          type: boolean
          description: Whether the scheduler is currently active
          example: true
        next_run_at:
          type: string
          format: date-time
          nullable: true
          description: Next scheduled crawl time (null if disabled)
        last_run_at:
          type: string
          format: date-time
          nullable: true

    SchedulerConfigUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        interval_minutes:
          type: integer
          minimum: 5
          maximum: 10080
          description: Crawl interval in minutes (5 min to 7 days)
        max_articles_per_summary:
          type: integer
          minimum: 10
          maximum: 500
          description: Max articles to include when generating a summary

    # ─── Settings ──────────────────────────────────────────
    Settings:
      type: object
      required: [llm, crawler]
      properties:
        llm:
          $ref: "#/components/schemas/LlmSettings"
        crawler:
          $ref: "#/components/schemas/CrawlerSettings"

    LlmSettings:
      type: object
      required: [provider, model, max_tokens, temperature]
      properties:
        provider:
          type: string
          enum: [claude, openai]
          example: "claude"
        model:
          type: string
          example: "claude-sonnet-4-20250514"
        claude_api_key_set:
          type: boolean
          description: Whether a Claude API key is configured (never returns the actual key)
          example: false
        openai_api_key_set:
          type: boolean
          description: Whether an OpenAI API key is configured (never returns the actual key)
          example: false
        max_tokens:
          type: integer
          minimum: 100
          maximum: 8000
          example: 2000
        temperature:
          type: number
          minimum: 0
          maximum: 1
          example: 0.3

    CrawlerSettings:
      type: object
      properties:
        request_timeout_seconds:
          type: integer
          minimum: 5
          maximum: 120
          example: 30
        max_content_snippet_length:
          type: integer
          minimum: 100
          maximum: 2000
          example: 500

    SettingsUpdate:
      type: object
      properties:
        llm:
          $ref: "#/components/schemas/LlmSettingsUpdate"
        crawler:
          $ref: "#/components/schemas/CrawlerSettings"

    LlmSettingsUpdate:
      type: object
      properties:
        provider:
          type: string
          enum: [claude, openai]
        model:
          type: string
        claude_api_key:
          type: string
          description: Set the Claude API key (write-only, never returned in GET)
        openai_api_key:
          type: string
          description: Set the OpenAI API key (write-only, never returned in GET)
        max_tokens:
          type: integer
          minimum: 100
          maximum: 8000
        temperature:
          type: number
          minimum: 0
          maximum: 1

    # ─── Common ────────────────────────────────────────────
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Source not found"